Replace crate: paste with pastey
`paste` is no longer maintained and thus triggers some advisory warnings. `pastey`is a drop-in replacement with the same API.

Unfortunately that alone is not enough to get rid of the advisory as various other libraries have a transitive dependency, such as:
- `simba`
- `rmp`
- `egui_dock`

But maybe a small step in the right direction.